feat(pst type): corrects wrong type PE-1918

+++ lib/pst/pst_contract_data.dart
@@ -87,7 +87,7 @@ class CommunityContractSettings {
   final List<String> communityDiscussionLinks;
   final String communityDescription;
   final ArweaveAddressType communityLogo;
-  final int fee;
+  final double fee;
 }
 
 class VaultItem {
